refactor(session-store): extract SessionListState sub-store (unit 1)
First class-decomposition unit per ADR-0002. Moves the session-list slice
(cold list, by-id/by-project indexes, live-sync + palette reference arrays,
per-project scan flags) and its mutators (setSessions, addSession,
updateSession, open-snapshot list-apply, scan flags) into a composed
SessionListState sub-store. SessionStore holds one instance and delegates;
private getter accessors keep the same live instances so the white-box
invariant tests in session-store-rename pass unchanged. Pure SessionCold
reconstructors (sessionColdFromExistingSession/WithMutableUpdates) moved to
session-cold-index alongside the other cold helpers.

session-store.svelte.ts 3997 -> 3876 LOC. check 0; rename 8/8; full store
vitest parity (7 pre-existing env/mock-infra fails, identical on baseline).
